--- ilogtail-controller/controller/dataDB/CURRENT
@@ -1,1 +1,0 @@
-MANIFEST-000003
