+++ ilogtail-controller/controller/tool/is_contain.go
@@ -1,0 +1,10 @@
+package tool
+
+func IsContain(items []interface{}, item interface{}) bool {
+	for _, eachItem := range items {
+		if eachItem == item {
+			return true
+		}
+	}
+	return false
+}
